feat(client):create usePluginCatalog
