+++ packages/amplication-client/src/Plugins/InstalledPluginSettings.tsx
@@ -24,7 +24,8 @@ import { Button, EnumButtonStyle } from "../Components/Button";
 import { AppContext } from "../context/appContext";
 import { AppRouteProps } from "../routes/routesUtil";
 import { formatError } from "../util/error";
-import usePlugins, { PluginVersion } from "./hooks/usePlugins";
+import usePlugins from "./hooks/usePlugins";
+import { PluginVersion } from "./hooks/usePluginCatalog";
 import "./InstalledPluginSettings.scss";
 import { PluginLogo } from "./PluginLogo";
