+++ packages/amplication-client/src/Plugins/InstalledPlugins.tsx
@@ -5,7 +5,8 @@ import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import { AppRouteProps } from "../routes/routesUtil";
 import { formatError } from "../util/error";
-import usePlugins, { Plugin } from "./hooks/usePlugins";
+import usePlugins from "./hooks/usePlugins";
+import { Plugin } from "./hooks/usePluginCatalog";
 import * as models from "../models";
 import PluginsCatalogItem from "./PluginsCatalogItem";
 import { EnumImages } from "../Components/SvgThemeImage";
